requirements.txt
asgiref==3.4.1
blis==0.7.4
catalogue==2.0.6
certifi==2020.12.5
chardet==4.0.0
ChatterBot==1.1.0
click==7.1.2
cymem==2.0.5
dj-database-url==0.5.0
Django==3.2.2
django-heroku==0.3.1
djangorestframework==3.12.4
en-core-web-lg==3.1.0
en-core-web-sm==2.1.0
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.5
mysql-connector-python==8.0.23
mysqlclient==2.0.3
nltk==3.5
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pathy==0.6.0
Pint==0.16.1
plac==0.9.6
preshed==3.0.5
protobuf==3.14.0
psycopg2==2.9.1
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.0
scipy==1.7.1
six==1.15.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.8
SQLAlchemy==1.3.23
sqlparse==0.4.1
srsly==2.4.1
thinc==8.0.9
threadpoolctl==2.2.0
tqdm==4.56.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
whitenoise==5.2.0
zipp==3.4.0
